+++ R/css_class.R
@@ -44,7 +44,7 @@
 #' }
 add_class <- function(ids, class_name) {
   send_message(
-    type = "toggleClass",
+    type = "dqToggleClass",
     ids = unname(ids),
     className = class_name,
     state = TRUE
@@ -55,21 +55,25 @@ add_class <- function(ids, class_name) {
 #' @rdname add_class
 remove_class <- function(ids, class_name) {
   send_message(
-    type = "toggleClass",
+    type = "dqToggleClass",
     ids = unname(ids),
     className = class_name,
     state = FALSE
   )
 }
 
-#' @param condition Condition to use for toggling the class.
+#' @param condition Condition to use for toggling the class, can be of length
+#' one or length of ids.
 #' @export
 #' @rdname add_class
 toggle_class <- function(ids, class_name, condition = NULL) {
-  send_message(
-    type = "toggleClass",
-    ids = unname(ids),
-    className = class_name,
-    state = condition
-  )
+  if (length(condition) > 0L) condition <- rep_len(condition, length(ids))
+  lapply(seq_along(ids), function(i) {
+    send_message(
+      type = "dqToggleClass",
+      ids = unname(ids[i]),
+      className = class_name,
+      state = condition[i]
+    )
+  })
 }