+++ R/dq_box.R
@@ -65,13 +65,11 @@ dq_box <- function(
   collapsible = FALSE, collapsed = FALSE, open_callback = FALSE
 ) {
 
-  if (is.null(id)) {
-    id <- paste0(sample(letters, 5L), collapse = "")
-  }
+  if (is.null(id)) id <- paste0("box-", random_id())
 
   box_class <- "dq-box"
   box_on_click <- NULL
-  body_id <- paste0(sample(letters, 8L), collapse = "")
+  body_id <- paste0("body-", random_id())
   head_class <- "dq-box-header clearfix"
   box_styles <- ""
 
@@ -172,6 +170,6 @@ create_box_title <- function(title, color) {
 #' @rdname dq_box
 update_dq_box <- function(id, collapsed = NULL, silent = FALSE) {
   send_message(
-    type = "updateBox", ids = id, collapsed = collapsed, silent = silent
+    type = "dqUpdateBox", ids = id, collapsed = collapsed, silent = silent
   )
 }