+++ R/state.R
@@ -44,19 +44,24 @@
 #'
 #' }
 enable <- function(ids) {
-  send_message(type = "toggleState", ids = ids, state = FALSE)
+  send_message(type = "dqToggleState", ids = ids, state = FALSE)
 }
 
 #' @export
 #' @rdname enable
 disable <- function(ids) {
-  send_message(type = "toggleState", ids = ids, state = TRUE)
+  send_message(type = "dqToggleState", ids = ids, state = TRUE)
 }
 
 #' @param condition Condition to be used for toggling the state (TRUE =
-#' disabled).
+#' disabled), can be of length one or length of ids.
 #' @export
 #' @rdname enable
 toggle_state <- function(ids, condition = NULL) {
-  send_message(type = "toggleState", ids = ids, state = condition)
+  if (length(condition) > 0L) condition <- rep_len(condition, length(ids))
+  lapply(seq_along(ids), function(i) send_message(
+    type = "dqToggleState",
+    ids = unname(ids[i]),
+    state = condition[i]
+  ))
 }