+++ inst/www/js/autocomplete-binding.js
@@ -22,8 +22,9 @@ $.extend(autocompleteBinding, {
   setValue: function (el, value) {
     var arr = $(el).data("options"),
       keys = Object.keys(arr),
-      labeled = !arr.length;
-    if (!labeled && arr.indexOf(value) > -1) {
+      labeled = !arr.length,
+      upd = $(el).data("create");
+    if (upd || (!labeled && arr.indexOf(value) > -1)) {
       $(el).attr("result", value);
       el.value = value;
     } else if (labeled && keys.indexOf(value) > -1) {
@@ -45,11 +46,9 @@ $.extend(autocompleteBinding, {
       callback(false);
     });
     $("body").on("click.autocompleteBinding", ".auto_selector", function (event) {
-      if (this.parentNode.id.startsWith(el.id)) {
-        var val = this.getElementsByTagName("input")[0].value;
-        setVal(el, val);
-        callback(false);
-      }
+      var val = this.getAttribute("data-value");
+      setVal(el, val);
+      callback(false);
     });
   },
 
@@ -59,7 +58,7 @@ $.extend(autocompleteBinding, {
 
   // Receive messages from the server.
   receiveMessage: function (el, data) {
-    if (data.hasOwnProperty("value")) el.value = data.value;
+    if (data.hasOwnProperty("value")) this.setValue(el, data.value);
 
     if (data.hasOwnProperty("label"))
       $(el).parent().find('label[for="' + el.id + '"]').text(data.label);
@@ -68,6 +67,7 @@ $.extend(autocompleteBinding, {
     if (data.hasOwnProperty("maxOptions")) $(el).data("max", data.maxOptions);
     if (data.hasOwnProperty("hideValues")) $(el).data("hide", data.hideValues);
     if (data.hasOwnProperty("placeholder")) el.placeholder = data.placeholder;
+    if (data.hasOwnProperty("create")) $(el).data("create", data.create);
 
     $(el).trigger("change");
   },
@@ -147,25 +147,25 @@ function autocomplete(inp) {
         if (labeled && !hideValues) {
           b.innerHTML += "<small>" + id + "</small>";
         }
+        b.setAttribute("data-value", id);
         a.appendChild(b);
         if (maxCount && ++count >= maxCount) break;
       }
     }
   });
   $(inp).on("keydown.autocompleteBinding", function (e) {
-    var x,
-      parent = document.getElementById(this.id + "autocomplete-list");
+    var x, parent = document.getElementById(this.id + "autocomplete-list");
     if (parent) x = parent.getElementsByTagName("div");
     if (e.keyCode == 40) {
-      /*arrow DOWN*/
+      //arrow DOWN
       currentFocus++;
       addActive(x);
     } else if (e.keyCode == 38) {
-      /*arrow UP*/
+      //arrow UP
       currentFocus--;
       addActive(x);
     } else if (e.keyCode == 13) {
-      /*ENTER key*/
+      //ENTER key
       e.preventDefault();
       if (currentFocus > -1) {
         if (x) x[currentFocus].click();